Remove usage of windows typedefs in registry code.

+++ src/Include/Registry.h
@@ -17,13 +17,12 @@
 #pragma once
 
 #include <string>
-#include <windows.h>
 
 namespace Registry
 {
 	auto ReadString(const std::string& subKey, const std::string& item) -> std::string;
 	auto SetString(const std::string& subKey, const std::string& item, const std::string& value) -> bool;
 
-	auto ReadDword(const std::string& subKey, const std::string& item) -> UINT;
-	auto SetDword(const std::string& subKey, const std::string& item, UINT value) -> bool;
+	auto ReadDword(const std::string& subKey, const std::string& item) -> unsigned long;
+	auto SetDword(const std::string& subKey, const std::string& item, unsigned long value) -> bool;
 }
